requirements.txt
argcomplete==3.6.3
asgiref==3.10.0
black==25.9.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
commitizen==4.9.1
coverage==7.11.0
decli==0.6.3
Deprecated==1.3.1
distlib==0.4.0
Django==5.2.8
django-environ==0.12.0
djangorestframework==3.16.1
filelock==3.20.0
flake8==7.3.0
identify==2.6.15
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-decouple==3.8
pytokens==0.3.0
PyYAML==6.0.3
questionary==2.1.1
sqlparse==0.5.3
termcolor==3.2.0
tomlkit==0.13.3
virtualenv==20.35.4
wcwidth==0.2.14
wrapt==2.0.1
django-autoslug==1.9.9
djangorestframework_simplejwt==5.5.1
django-filter==25.2
